+++ packages/layout/CHANGELOG.md
@@ -7,9 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+
+- vite warning on inline css import resolved
+- json schema import fixed for `mlc-config` web component
+
 ### Versioning
 
+- Storybook updated
 - Development dependencies updated
+- Typescript set to `^4.9.4` with yarn `3.3.1`
 - Typescript is set to `4.9.3` due to yarn berry temporary incompatibility
 - Direct dependencies `@ctrl/tinycolor` was updated
 - Direct dependencies `lit-html` was updated and deduped due to this [issue](https://github.com/lit/lit/issues/3241)